Разобъёт util.js на несколько файлов, в которых функции объединены по смыслу, и разместит эти файлы в новой директории utils. Поправит импорты.
